moves survivor repo to internal
