+++ api/survivor/internal/repo.go
@@ -1,4 +1,4 @@
-package survivor
+package internal
 
 import (
 	"context"
@@ -7,7 +7,7 @@ import (
 	"log"
 )
 
-type postgresRepo struct {
+type PostGresRepo struct {
 	pool store.Connection
 }
 
@@ -34,20 +34,22 @@ type Survivor struct {
 	Understanding    int
 }
 
-func newRepo(d store.Connection) *postgresRepo {
-	return &postgresRepo{pool: d}
+func NewRepo(d store.Connection) *PostGresRepo {
+	return &PostGresRepo{pool: d}
 }
 
-func getAllSurvivorsForSettlement(settlementId int) string {
-	return fmt.Sprintf("SELECT * FROM campaign.survivor WHERE settlement = %d", settlementId)
+func (r PostGresRepo) GetAllSurvivorsForSettlement(ctx context.Context, settlementId int) ([]Survivor, error) {
+	query := fmt.Sprintf("SELECT * FROM campaign.survivor WHERE settlement = %d", settlementId)
+	survivors, err := r.find(ctx, query)
+	return survivors, err
 }
 
-func (r postgresRepo) Find(ctx context.Context, query string) ([]Survivor, error) {
+func (r PostGresRepo) find(ctx context.Context, query string) ([]Survivor, error) {
 	log.Default().Println(query)
-	rows, err := r.pool.Query(ctx, query)
-	if err != nil {
-		log.Default().Println(err.Error())
-		return []Survivor{}, err
+	rows, queryErr := r.pool.Query(ctx, query)
+	if queryErr != nil {
+		log.Default().Println(queryErr.Error())
+		return nil, queryErr
 	}
 	defer rows.Close()
 	survivors := []Survivor{}
